Moved all utility modules and tested the data collection script successfully
